refactor: extract messenger_curated_communities.go
